Add basic support for Atmel SAMW25 Xplained Pro board (but no WIFi yet) Make graphics prims return an object. Probably not necessary, but theoretically more "correct". Tweak unit32 typedef to be consistent with WiFi101 library (for later).

+++ vm/mem.h
@@ -15,7 +15,7 @@ extern "C" {
 
 typedef unsigned char uint8;
 typedef unsigned short uint16;
-typedef unsigned int uint32;
+typedef unsigned long uint32;
 
 // Boolean constants for readability
